+++ python/cudf/cudf/io/parquet.py
@@ -237,7 +237,7 @@ def to_parquet(
         if partition_cols:
             write_to_dataset(
                 df,
-                path,
+                root_path=path,
                 partition_cols=partition_cols,
                 preserve_index=index,
                 **kwargs,
@@ -254,8 +254,8 @@ def to_parquet(
 
         return libparquet.write_parquet(
             df,
-            path,
-            index,
+            path=path,
+            index=index,
             compression=compression,
             statistics=statistics,
             metadata_file_path=metadata_file_path,
@@ -268,7 +268,11 @@ def to_parquet(
 
         pa_table = df.to_arrow(preserve_index=index)
         return pq.write_to_dataset(
-            pa_table, path, partition_cols=partition_cols, *args, **kwargs
+            pa_table,
+            root_path=path,
+            partition_cols=partition_cols,
+            *args,
+            **kwargs,
         )