+++ python/cudf/cudf/tests/utils.py
@@ -5,8 +5,8 @@ import numpy as np
 import pandas as pd
 import pandas.util.testing as tm
 
+import cudf
 import cudf.utils.dtypes as dtypeutils
-from cudf import Series
 from cudf._lib.null_mask import bitmask_allocation_size_bytes
 
 supported_numpy_dtypes = [
@@ -59,7 +59,7 @@ def count_zero(arr):
     return np.count_nonzero(arr == 0)
 
 
-def assert_eq(a, b, **kwargs):
+def assert_eq(left, right, **kwargs):
     """ Assert that two cudf-like things are equivalent
 
     This equality test works for pandas/cudf dataframes/series/indexes/scalars
@@ -69,43 +69,43 @@ def assert_eq(a, b, **kwargs):
     """
     __tracebackhide__ = True
 
-    if hasattr(a, "to_pandas"):
-        a = a.to_pandas()
-    if hasattr(b, "to_pandas"):
-        b = b.to_pandas()
-    if isinstance(a, cupy.ndarray):
-        a = cupy.asnumpy(a)
-    if isinstance(b, cupy.ndarray):
-        b = cupy.asnumpy(b)
+    if hasattr(left, "to_pandas"):
+        left = left.to_pandas()
+    if hasattr(right, "to_pandas"):
+        right = right.to_pandas()
+    if isinstance(left, cupy.ndarray):
+        left = cupy.asnumpy(left)
+    if isinstance(right, cupy.ndarray):
+        right = cupy.asnumpy(right)
 
-    if isinstance(a, pd.DataFrame):
-        tm.assert_frame_equal(a, b, **kwargs)
-    elif isinstance(a, pd.Series):
-        tm.assert_series_equal(a, b, **kwargs)
-    elif isinstance(a, pd.Index):
-        tm.assert_index_equal(a, b, **kwargs)
-    elif isinstance(a, np.ndarray) and isinstance(b, np.ndarray):
-        if np.issubdtype(a.dtype, np.floating) and np.issubdtype(
-            b.dtype, np.floating
+    if isinstance(left, pd.DataFrame):
+        tm.assert_frame_equal(left, right, **kwargs)
+    elif isinstance(left, pd.Series):
+        tm.assert_series_equal(left, right, **kwargs)
+    elif isinstance(left, pd.Index):
+        tm.assert_index_equal(left, right, **kwargs)
+    elif isinstance(left, np.ndarray) and isinstance(right, np.ndarray):
+        if np.issubdtype(left.dtype, np.floating) and np.issubdtype(
+            right.dtype, np.floating
         ):
-            assert np.allclose(a, b, equal_nan=True)
+            assert np.allclose(left, right, equal_nan=True)
         else:
-            assert np.array_equal(a, b)
+            assert np.array_equal(left, right)
     else:
-        if a == b:
+        if left == right:
             return True
         else:
-            if np.isnan(a):
-                assert np.isnan(b)
+            if np.isnan(left):
+                assert np.isnan(right)
             else:
-                assert np.allclose(a, b, equal_nan=True)
+                assert np.allclose(left, right, equal_nan=True)
     return True
 
 
-def assert_neq(a, b, **kwargs):
+def assert_neq(left, right, **kwargs):
     __tracebackhide__ = True
     try:
-        assert_eq(a, b, **kwargs)
+        assert_eq(left, right, **kwargs)
     except AssertionError:
         pass
     else:
@@ -146,9 +146,9 @@ def gen_rand(dtype, size, **kwargs):
 def gen_rand_series(dtype, size, **kwargs):
     values = gen_rand(dtype, size, **kwargs)
     if kwargs.get("has_nulls", False):
-        return Series.from_masked_array(values, random_bitmask(size))
+        return cudf.Series.from_masked_array(values, random_bitmask(size))
 
-    return Series(values)
+    return cudf.Series(values)
 
 
 @contextmanager