+++ python/custreamz/custreamz/tests/test_dataframes.py
@@ -444,7 +444,9 @@ def test_setitem_overwrites(stream):
 def test_rolling_count_aggregations(
     op, window, m, pre_get, post_get, kwargs, stream
 ):
-    index = pd.DatetimeIndex(start="2000-01-01", end="2000-01-03", freq="1h")
+    index = pd.DatetimeIndex(
+        pd.date_range("2000-01-01", "2000-01-03", freq="1h")
+    )
     df = cudf.DataFrame({"x": np.arange(len(index))}, index=index)
 
     expected = getattr(post_get(pre_get(df).rolling(window)), op)(**kwargs)
@@ -642,7 +644,9 @@ def test_windowing_n(func, n, getter):
     "indexer", [lambda g: g, lambda g: g[["x"]], lambda g: g[["x", "y"]]]
 )
 def test_groupby_windowing_value(func, value, getter, grouper, indexer):
-    index = pd.DatetimeIndex(start="2000-01-01", end="2000-01-03", freq="1h")
+    index = pd.DatetimeIndex(
+        pd.date_range("2000-01-01", "2000-01-03", freq="1h")
+    )
     df = cudf.DataFrame(
         {
             "x": np.arange(len(index), dtype=float),